C++11 Fixes: Part 2.

+++ cpu_ref/rsCpuIntrinsicInlines.h
@@ -64,7 +64,8 @@ static inline uint4 convert_uint4(uchar4 i) {
 }
 
 static inline int4 convert_int4(float4 i) {
-    int4 f4 = {i.x, i.y, i.z, i.w};
+    int4 f4 = {static_cast<int>(i.x), static_cast<int>(i.y),
+               static_cast<int>(i.z), static_cast<int>(i.w)};
     return f4;
 }
